+++ CommHandler/CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [unreleased]
 
 ## Added
 
 - client for SGr declaration library
+- JSONata expressions as response query in REST / messaging data points and message filters
 
 ### Changed